requirements.txt
python-dotenv
mne
scikit-learn
matplotlib
tensorflow==2.5